Added spreadsheet that contains all Hadoop benchmark data for Cloud OneFS
This spreadsheet contains Hadoop benchmark data for 56 x tier 1 Cloud OneFS nodes deployed in the us-east4 region of Google Cloud.  At the time of running this benchmark, only half the network connections were enabled on the Cloud OneFS tier in this region.  Even under these circumstances, Cloud OneFS executed TeraGen in 1/2 the time of running the same test with dataproc using google cloud storage and 1/3 the time with TeraSort.